mlx-opt: full overnight evolve results (5 levels + batch sweep + length curve)
RESULTS — M3 8GB, MLX 0.32, L=140, 3-warmup 5-run:

  Level            ms      Speedup   MAE     Active mem
  baseline fp32   56.5ms    1.00x   0.000    1332 MB
  mx.compile      55.5ms    1.02x   0.000    1332 MB
  bf16 only       48.2ms    1.17x   1.212     667 MB
  bf16+compile    47.3ms    1.19x   1.212     667 MB
  fused_ln        47.1ms    1.20x   1.596     667 MB

BATCH SWEEP (L=140, opt=3):
  B=1   48.4ms  20.6 seq/s  1.00x
  B=4  139.2ms  28.7 seq/s  1.39x
  B=8  267.5ms  29.9 seq/s  1.45x
  B=16 517.3ms  30.9 seq/s  1.50x

LENGTH CURVE opt=3: 2072 tok/s at L=50 → 3811 tok/s at L=500

KEY FINDINGS:
- bf16+compile is the recommended setting (1.19x, MAE < 1.5)
- Batching to B=8 adds 45% throughput for ProteinGym scoring
- bf16 without compile: needs JIT to be faster (BF16 dispatch overhead)
- Per-layer: first 15 blocks 2.83ms, last 15 blocks 2.02ms (decreasing cost)
- Fix: bf16 log_softmax must cast to float32 before numpy conversion

+++ scoring/esmc/mlx_esmc.py
@@ -360,7 +360,7 @@ class EsmcMLX(nn.Module):
             logits = self(ids_mx)                           # [B, L, 64]
             mx.eval(logits)
 
-            log_probs = np.array(nn.log_softmax(logits, axis=-1))  # [B, L, 64]
+            log_probs = np.array(nn.log_softmax(logits.astype(mx.float32), axis=-1))  # [B, L, 64]
             for b_idx, pos in enumerate(batch_pos):
                 pos_to_lp[pos] = log_probs[b_idx, pos + 1]  # log-probs at masked pos
